# Docs/AssessmentServices.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Default module",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/customers/": {
      "post": {
        "summary": "Add Customer",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "example": "",
                    "type": "string",
                    "format": "binary"
                  },
                  "name": {
                    "example": "Mohammed Salah",
                    "type": "string"
                  },
                  "phone": {
                    "example": "0545006176",
                    "type": "string"
                  }
                }
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "summary": "Get All Customers",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Assessment"
    }
  ],
  "security": []
}